Move activation hooks to a separate file in "includes" folder
